meeting api

+++ src/main/java/pl/umk/mat/planner/meeting/Meeting.java
@@ -4,9 +4,7 @@ import pl.umk.mat.planner.event.Event;
 
 import javax.persistence.*;
 import java.time.OffsetDateTime;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Table(name = "meeting")
@@ -14,7 +12,7 @@ public class Meeting {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", nullable = false, updatable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "date_from", nullable = false)
     private OffsetDateTime dateFrom;
@@ -34,7 +32,7 @@ public class Meeting {
         this.events = events;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
